Move RETURN_TO_SETUP_GUIDE_PARAM into embedding
