+++ frontend/src/metabase/admin/databases/hooks/use-database-connection.ts
@@ -1,9 +1,9 @@
 import { push } from "react-router-redux";
 import { t } from "ttag";
 
-import { RETURN_TO_SETUP_GUIDE_PARAM } from "metabase/admin/embedding/embedding-hub/constants";
 import { skipToken, useGetDatabaseQuery } from "metabase/api";
 import { getDefaultEngineKey } from "metabase/databases/utils/engine";
+import { RETURN_TO_SETUP_GUIDE_PARAM } from "metabase/embedding/constants";
 import { PLUGIN_DB_ROUTING } from "metabase/plugins";
 import { useDispatch } from "metabase/redux";
 import type { DatabaseId, Engine, EngineKey } from "metabase-types/api";
