+++ frontend/src/metabase/admin/embedding/embedding-hub/components/EmbeddingHubXrayPickerModal.tsx
@@ -1,8 +1,8 @@
 import { push } from "react-router-redux";
 import { t } from "ttag";
 
-import { RETURN_TO_SETUP_GUIDE_PARAM } from "metabase/admin/embedding/embedding-hub/constants";
 import { DataPickerModal } from "metabase/common/components/Pickers/DataPicker";
+import { RETURN_TO_SETUP_GUIDE_PARAM } from "metabase/embedding/constants";
 import { useDispatch } from "metabase/redux";
 import type { TableId } from "metabase-types/api";
