--- frontend/src/metabase/admin/embedding/embedding-hub/constants.ts
@@ -1,2 +1,0 @@
-export const EMBEDDING_SETUP_GUIDE_PATH = "/admin/embedding/setup-guide";
-export const RETURN_TO_SETUP_GUIDE_PARAM = "returnToEmbeddingSetupGuide";
